Experiment with gcovr generation of cobertura XML
Making it work with sandboxing requires patches to gcovr. By default, gcovr
looks for error messages from gcov, and skips files if it cannot resolve
the source file. We don't have the source files available in the sandbox, but
the source files aren't actually needed for XML generation.

One possible workaround is to create empty source files using the manifest (I
think we did that with lcov until I found another workaround). The problem is
that gcovr still skips coverage files if it can't find even a single source
file, and the manifest intentionally only contains those that are supposed to
be covered.

The alternative is to patch gcovr.