Rewrite the remote coverage tools repository
Use a bzl rule generating a simple "java -jar" wrapper script instead of a java_binary rule. The wrapper script generated by java_binary requires a runfiles tree, which does not work for the coverage tools when executed remotely. In addition, the wrapper script is unnecessarily smart when the binary is already based on a complete deploy jar.

Additionally, making the runfiles tree available for the coverage tools is not straightforward where the test action is generated. We neither want to pollute the normal test runfiles tree nor do we want all Starlark test rules to have to account for additional runfiles.

Instead, we create a bzl rule that declares the deploy jar and a simple wrapper script as files_to_build (from which files_to_run gets derived) and uses a direct invocation of "java", which it assumes to be present on the remote machine.

Unfortunately, this does not work on Windows or other platforms that do not have "/bin/bash" available, but it may serve as a workaround for users who currently can't use coverage with remote execution, e.g., by setting --override_repository.

Change-Id: I66b32b2c335fe930da2376311e6f3c07a50ec84e