Analysis: access OS.getCurrent through CoreOptions
This helped me track down the places where the current OS (where Bazel runs) is
used during analysis. Unfortunately, it doesn't address the elephant in the
room, file paths, which also behave differently depending on the current OS.

Not sure whether it's useful to merge this before figuring out a plan for
paths.

Change-Id: I20dff1fb332d4567b621d65e43b05a00905f75b0