DynamicExecutionModule: disable SpawnCache
When dynamic execution is active, disable the spawn cache. The RemoteModule
always provides a remote spawn cache when remote execution is active. However,
when dynamic execution is active, we don't want to use remote execution *and*
local execution with remote caching.

In fact, trying to use both leads to an assertion error in the new dynamic
execution implementation.

The downside of this change is that it no longer allows us to use the remote
cache for actions that disallow remote execution. It also doesn't allow using
the disk cache for the local branch (although that is also currently not
allowed in the RemoteModule).

Fixes #12364.

Change-Id: I5aa192143604b01a65d68ac73f1d363957503396