RemoteActionFS: add --remote_fetch_symlink_inputs
The RemoteActionFS is currently used if --remote_download_minimal is
enabled. In that case, Bazel is supposed to minimize downloads from a
remote execution service.

Before this change, Bazel downloaded files in order to create local
symlinks, but despite the documentation this does not seem to be
necessary.

I tested with the following scenario that is specifically designed to
trigger this case:

// Create an executable file (remote)
genrule(
  name = "genhello",
  outs = ["hello.sh"],
  cmd = "echo 'echo Hello World!' > $@",
  executable = True,
)

// Symlink the executable file in a different location (local)
sh_binary(
  name = "hello",
  srcs = ["hello.sh"],
)

// Execute the file from the symlinked location (remote)
genrule(
  name = "continued",
  outs = ["continued.txt"],
  tools = [":hello"],
  cmd = "$(location :hello) > $@",
)

Change-Id: I485c56740003f6a32bf416ae58746ee7614647b4